requirements.txt
bandit==1.7.9
black==24.8.0
click==8.1.7
fastapi==0.115.0
httpx==0.27.0
mypy==1.11.2
PyJWT==2.9.0
pydantic==2.8.0
pytest==8.2.0
pytest-asyncio==0.23.0
pytest-cov==5.0.0
PyYAML==6.0.1
ruff==0.6.5
slowapi==0.1.9
types-PyYAML==6.0.12.20240917
uvicorn==0.30.0
websockets>=12.0
